Does Typing Only Pytest work for you?
It does for me. If so, it might be nice to simplify the command for testing this project to simply 'pytest`.

Also, when you run pytest do you get Deprecation Warnings that seem to come from something upstream in bs4? We could try to fix that upstream. It looks simple to fix. (Famous last words.)